Improve module loading
Modules are now loaded by a simple script virtme-loadmods.  We still
try to use udev, but it's less important than before.

virtme-loadmods isn't great: it will not discover modaliases that
are enumerated after it exits.  It's still a lot better than
nothing, and there may be demand for a udev-free mode at some point.

Signed-off-by: Andy Lutomirski <luto@amacapital.net>