virtme-init: make resolv.conf world-readable.
'mktemp' creates files that are only accessible to the current user
(and doesn't have any options to change this), but in this case the
file we're creating is going to end up as resolv.conf, which should be
world-readable.  So here we manually chmod it as such after creating it.

Signed-off-by: Zev Weiss <zev@bewilderbeest.net>