virtme-run: provide a way to support modules with --kdir
When --kdir is provided there's not an easy and automated way to use
kernel modules from the guest.

A possible solution could be to install all the modules inside the
kernel build directory (argument of --kdir) inside a temporary directory
.tmp_moddir/ as following:

 $ make modules_install INSTALL_MOD_DIR=.tmp_moddir

In this way if .tmp_moddir is found the guest can use the kernel modules
from this directory, otherwise fallback to the old behavior and do not
provide an automated way to access the kernel modules.

TODO: document this .tmp_moddir/ directory somewhere.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>