virtme/architectures.py: Fix ppc64le support
Fix virtme usage on ppc64le. virtme-run tries to match the uname.machine
with all supported architectures and returns an Arch object. There
isn't a 'ppc64le' version, so it ends up in the Arch_unknown('ppc64le'),
which later on is used to find the qemu-system-$arch command.

There isn't a qemu-system-ppc64le, only qemu-system-{ppc,ppc64}.
Replicating the configuration for ppc64 and ppc64le fixes the problem.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>