run.py: append -kc to kernel version
Append a '-kc' suffix to the kernel version for all the kernels built
using KernelCraft.

This provides a better compatibility for external modules (especially
when the kernel is built on a remote host), since it avoids having
different local versions appended.

Moreover, it makes easier to distinguish between the real host kernel
and the kernel running inside KernelCraft.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>