virtme-init: fix resolv.conf symlink path
In some distro /etc/resolv.conf is a symlink to a systemd-generated file
into /run/systemd/, like /run/systemd/resolve/stub-resolv.conf.

However the symlink is actually pointing to
../run/systemd/resolve/stub-resolv.conf.

When virtme_chdir is used we end up creating a symlink in the wrong
path, so make sure to resolve the symlink from the root of the
filesystem and not from the current working directory.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>