run: Add new --compiler argument
There are cases when compiling a kernel with a more recent compiler can
return errors. For example, on curent openSUSE Tumbleweed, gcc-13 is
present, and it errors out on a couple of places.

For such cases, --compiler would help to test using different compiler
versions, like in the example, using gcc-11 can compile the mainline
kernel without problems.

Signed-off-by: Marcos Paulo de Souza <mpdesouza@suse.com>