virtme-configkernel: enable BLK_DEV_INITRD as optional
We need to enable BLK_DEV_INITRD to support virtme-run --force-initramfs
that would be broken otherwise.

However, we may also want to disable it if we use the default
initramfs-less boot, so make it optional, in this way we can decide to
override this config and disable it passing a custom config snippet
(using virtme-configkernel --custom FILE).

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>