run: riscv64: print cmdline size warning without blocking
Do not wait for input when printing the command line size warning for
riscv64. It is enough to just print the warning, waiting for user input
can be annoying when using KernelCraft in an automated environment.

Moreover, move the warning just before starting virtme, so in case of
errors the warning can be noticed easily.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>