virtme: support virtio-fs
Virtiofs is a shared file system that lets virtual machines access a
directory tree on the host. Unlike existing approaches, it is designed
to offer local file system semantics and performance [1].

Compared to the 9P filesystem, virtio-fs offers consistent performance
advantages by delivering access times that closely approach the native
performance of the host filesystem.

Test case:
 - Run `git diff` from a linux git repo

Results:
       9p-fs: 283.814s
   virtio-fs:   1.174s

Speed up in this particular case is approximately 241.96 times faster.

Virtme will try to automatically enable the new virtio-fs support: if
virtiofsd is found in the $PATH or in /usr/libexec, it'll be used to
export the host's rootfs and qemu options will be automatically adjusted
to use it. Otherwise virtme will fall-back to the legacy 9p-fs mode.

Moreover, introduce the new virtme-run option --force-9p to force the
legacy 9p-fs mode, even when a proper virtiofsd is detected in the
system.

Notes:

 - need to rely on the new Rust implementation of virtiofsd [2],
   instead of the daemon shipped in qemu, because the latter doesn't
   support unprivileged mode execution and it would be totally unsafe to
   export the whole rootfs of the host running as virtiofsd as root.

 - At the moment the Rust's virtiofsd doesn't seem to handle fs
   permissions correctly if used together with overlayfs [3], for this
   reason we need to apply a special patch to virtiofsd [4]. With this
   patch applied we can correctly export the entire root filesystem of
   the host inside the virtme guest and enable the full copy-on-write
   mode everywhere using overlayfs and tmpfs (maintaining the same
   behavior of the legacy 9p fs case).

 - Right now virtiofs is used only in interactive mode; virtiofs
   requires to define a chardev device in the qemu options that seems to
   conflict with the chardev device used for the virtio-console
   redirection (required to collect the script/command output from the
   guest to the host), so at the moment passing the option --script-exec
   or --script-sh would implicitly disable virtiofs.

[1] https://virtio-fs.gitlab.io/
[2] https://gitlab.com/virtio-fs/virtiofsd
[3] https://gitlab.com/virtio-fs/virtiofsd/-/issues/99
[4] https://github.com/arighi/virtiofsd/commit/8c72edfce01e9f5415bf1b963ec4e70042e22231

Suggested-by: Andy Lutomirski <luto@kernel.org>
Signed-off-by: Andrea Righi <andrea.righi@canonical.com>