virtme-ng: drop --release option
Drop the --release option that doesn't really provide any real benefit.

This option was added when virtme-ng was strictly dependent on git, as a
helper to quickly add remotes, but it is actually faster to use git
directly to add remotes. Moreover, we would need to constantly maintain
remotes in the local virtme-ng configuration file, that is just extra
work for no real benefit. remotes in the local virtme-ng configuration
file, that is just extra work for no real benefit.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>