virtme-run: drop the 9p mount for virtme guest tools
There is no reason to mount the guest tools in a separate 9p filesystem,
since we provide the whole root to the guest we can simply use a bind
mount to relocate the guest tools in the proper location.

This also saves a little bit of time to start the virtme instance.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>