virtme: support 'microvm' virtual architecture
microvm is a minimalist QEMU machine type without PCI nor ACPI support,
designed for short-lived guests. microvm also establishes a baseline for
benchmarking and optimizing both QEMU and guest operating systems, since
it is optimized for both boot time and footprint [1].

This architecture also provides a better integration and support with
virtio-fs and in perspective it would allow in the future to potentially
drop ACPI and PCI in order to produce a more lightweight kernel.and PCI
in order to produce a more lightweight kernel.

This provides a solution for issue #10.

[1] https://qemu.readthedocs.io/en/latest/system/i386/microvm.html

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>