virtme-ng: introduce --build and make run the default
By default virtme-ng should try to run kernels, instead of building
them, to maintain the best level of compatibility with the legacy
virtme-run.

For this reason introduce the option --build that can be passed to build
a kernel in the current working directory before trying to execute it.

Moreover, get rid of the deprecated option --skip-build.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>