virtme: proper stdin/stdout/stderr redirection with scripts
Virtio serial ports are designed to support a single writer at a time,
which means that only one process or application can write to the port
at any given moment.

When using a virtio serial port, the EBUSY error can occur if multiple
writers are attempting to access the port simultaneously.

This can happen for example when a script running inside the guest tries
to write to /dev/stdout or /dev/stderr directly, that are mapped to the
same virtio serial port already opened to forward stdout and stderr to
the host.

To prevent this problem create two separate virtio serial ports to
handle writes directly to /dev/stdout and /dev/stderr that will be all
redirected to stdout on the host.

Moreover, create independent virtio serial ports for stdout and stderr,
so we can handle stdout/stderr separately, as if we were running the
script natively on the host.

Example:

 $ echo hello | virtme-ng -r --exec "cat -" | md5sum
 b1946ac92492d2347c6235b4d2611184  -
 $ echo hello | virtme-ng -r --exec "cat - >/dev/stderr" | md5sum
 hello
 d41d8cd98f00b204e9800998ecf8427e  -
 $ echo hello | virtme-ng -r --exec "cat - >/dev/stderr" 2>&1 | md5sum
 b1946ac92492d2347c6235b4d2611184  -

This fixes issue #8.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>