virtme-init: kill guest when main interactive session exits
It's a lot easier to simply exit from the main interactive session to
kill a virtme-ng session, rather than the "ctrl+a x" combo to force-exit
from the QEMU session.

Using a simple 'exit' makes the interface a lot more user-friendly and
more intuitive.

I don't see any particular downside about changing this behavior. We may
have regressions or potentially break previous workflow with this
change. If it becomes a problem we'll introduce an option to restore the
legacy behavior.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>