virtme-ng: allow to run graphical applications
Allow to run individual graphical applications inside a virtme-ng
session.

Change the `--graphics / -g` option to accept an optional argument. If
an argument is passed, it will be interpreted as a graphical application
to run inside a virtme-ng session.

virtme-ng will simply use `xinit` with a bare minimum `.xinitrc` to
start the application in a minimal X graphical environment.

This is the first prototype of the functionality mentioned in issue #15.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>