virtme-ng: drop crash dependency
Change --dump to always generate a memory dump file, instead of
attaching directly a session via the crash tool. This feature is pretty
much useless, since most of the time we actually want to generate a
separate file and inspect it later.

This also allows to drop the dependency of the 'crash' tool and still be
able to generate memory dump files.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>