virtme_ng_init: properly synchronize udev and network startup
There is a potential race between udev and network startup, the latter
could run too early, failing to detect a proper network interface and
failing to setup networking.

If that happens we need to wait a little bit to make sure udev has
registered corrected the virtio network device, and try again to setup
the network.

Moreover, refactor the networking code a little bit to make it more
readable.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>