virtme: prevent network interface renaming
Sometimes a network interface can be renamed as soon as we bring it up.
This causes a failure when we run udhcpc, because we evaluate the
interface name before bringing it up.

To prevent this issue always pass "net.ifnames=0 biosdevname=0" to the
kernel boot parameters, so the network interface won't be renamed during
the boot process.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>