virtme-ng: drop --quiet and provide --verbose
Running in quiet mode should be the default and we should provide a
`--verbose` option to increse console output verbosity.

Also change the logic of executing scripts: if `--verbose` is passed
then automatically enable `--show-boot-console` with virtme-run.

This makes everything a lot more intuitive and easier to debug potential
issues.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>