virtme-ng: try to automatically fix permissions of snapd state.json
Instead of showing a warning suggesting to run
`chmod +r /var/lib/snapd/state.json` when snaps are enabled we may try
to automatically give read access to state.json to the current user
using POSIX ACL (via getfacl / setfacl).

It's still potentially an unsafe operation in the system, so make sure
to print a big warning before changing any permission.

Even if it's still an ugly workaround to support snaps, it seems better
than asking users to run a `chmod +r` on a system file.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>