deb: add virtme-virtiofsd dependency
There are certain features in virtiofsd that are not upstream yet, but
virtme-ng relies on them.

For this reason temporarily provide a virtme-virtiofsd package for now,
to provide a custom virtiofsd with those fixes applied (in particular
the fix to support overlayfs on top of virtiofsd and the dependency that
breaks qemu).

This allows to have all the required components to speed up virtme-ng
instances without having to deal with upstream issues.

Once everything will be fixed upstream we will add back the dependency
on the upstream virtiofsd and drop virtme-virtiofsd.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>