virtme: init process refactoring
Move all the commands required to initialize the system into
virtme-ng-init (or virtme-init).

Moreover, when running scripts inside the guest don't use /run as a
temporary place to store the script that needs to be executed, use /mnt
instead, because /run needs to be re-mounted with tmpfs by
virtme-ng-init or virtme-init.

This is the first step to provide an initramfs-less way of executing
scripts inside the guest, that should help to improve boot time even
more (for the script execution case).

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>