virtme-ng: add virtiofsd as a git submodule
Include virtiofsd directly for now, so that we don't have to deal with
upstream dependencies (such as conflicts with qemu and missing features
that are required by virtme-ng).

The binary of virtiofsd will be included in the guest tool directory on
the host.

Once all these issues will be fixed upstream we will drop this
submodule and we can simply rely on the upstream virtiofsd.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>