setup.py: provide build options to exclude virtiofsd and virtme-ng-init
Allow to exclude virtiofsd and virtme-ng-init selectively from the
virtme guest binaries shipped in virtme-ng.

This can be done setting the following environment variables:

 - BUILD_VIRTIOFSD=1|0: ship virtiofsd with virtme-ng (default off)

 - BUILD_VIRTME_NG_INIT=1|0: ship virtme-ng-init with virtme-ng (default off)

Example:
  - install virtme-ng enabling embedded virtiofsd and virtme-ng-init:

 $ BUILD_VIRTIOFSD=1 BUILD_VIRTME_NG_INIT=1 pip install -r requirements.txt --verbose

Enabling them selectively can be helpful with packaging, because now we
have the options to embed everything in a single standalone virtme-ng
package or provide the optional components as external package
dependencies.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>