setup.py: introduce a .config to define build options
Allow to define build options (such as BUILD_VIRTIOFSD and
BUILD_VIRTME_NG) in a .config file that will be parsed by setup.py
during the build.

This can be used to decide which components should be enabled during the
build of the package.

With this change in place we can restore the old debian/rules and
debian/control, removing the cargo build hacks. The Rust components can
now be provided as external packages and managed independently, making
the packaging more suitable for Debian inclusion, or we can build a
all-in-one package (e.g., for the Ubuntu deb.), simply by changing the
settings in .config.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>