virtme-ng: encode shell command passed via --exec to base64
Commit 18680da ("virtme-ng: pass --exec command via /proc/cmdline")
introduced a regression found by the mutter test case [1].

Passing the --exec command via /proc/cmdline directly, instead of using
a temporary initramfs, is faster, but it has the downside that we don't
support special characters anymore (such as double quotes for example).

To solve this regression and avoid breaking the old behavior we can pass
the command as a base64-encoded string, that will be decoded in the
guest by virtme-ng-init/virtme-init.

The overhaed of the base64 encoding/decoding is negligible (definitely
better than creating a temporary initramfs) and we can use any special
character in the shell command.

This fixes issue #27.

[1] https://gitlab.gnome.org/GNOME/mutter/-/jobs/2917515

Fixes: 18680da ("virtme-ng: pass --exec command via /proc/cmdline")
Signed-off-by: Andrea Righi <andrea.righi@canonical.com>