virtme-ng: check for a statically linked busybox
In order to reduce the memory footprint of the initramfs used by
virtme-ng (and to avoid the hassle of figuring out the lib dependencies)
we rely on a statically build busybox binary, that is physically copied
from the host system into the initramfs (when the guest requires it).

However, some distro (e.g. Debian) provides a dynamically linked busybox
by default and when that is used in the initramfs we get quite obscure
errors on boot.

For this reason add an explicit check to determine if the busybox binary
detected in the system is statically linked and print an explicit error,
suggesting to install busybox-static if a statically linked version
cannot be found.

Also add busybox-static as a dependency for the deb package.

This fixes issue #29.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>