virtme: print a warning if "file" is not installed
We rely on `file` to determine the kernel version from a generic kernel
file (that is the most reliable way, especially with compressed
kernels).

However if `file` is not installed in the system or it's not available
in the PATH we may get an error. In this case print a warning and try to
run the kernel anyway (if the kernel version cannot be determined
modules may not work, but a warning will be printed and it's still
better than just failing with an ugly trace).

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>