virtme-ng: --run should default to current directory
Do not create an instance with the running kernel by default, but try to
run the kernel from the current working directory.

The typical use-case for virtme-ng is to run a kernel from the current
working directory, so make this behavior the default one if no other
argument is specified.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>