deb: drop deb packaging
Now that virtme-ng is available in Debian we don't need to maintain the
deb packaging anymore in this repo, we can simply rely on python
setuptools to provide a form of distro-agnostic packaging, then distros
will provide their own packaging scripts/configs.

This should also help to keep the deb packaging in sync between upstream
and Debian (and all the derivative distro, such as Ubuntu).

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>