virtme-ng: fix --user behavior
Properly implement the following behavior regarding the default user in
virtme-ng sessions:

 - default user for non-interactive sessions is root, unless --user is
   specified

 - default user for interactive sessions is the current user, unless
   --user is specified

We want to maintain this distinction to avoid breaking the old behavior,
but also for performance reasons: running non-interactive scripts as
root is faster, because we can get rid of the "user switch" step and
immediately run the target script as soon as the boot is done.

On the other hand, switching automatically to the target user during
interactive sessions has the benefit of giving the user the feeling of
running the target kernel inside the same exact environment of the host.

For the above reasons it makes sense to maintain this behavior.

However, we should also honor the `--user USER` when it's used and right
now we are honoring it for interactive sessions only. Fix this and make
sure the --user option is considered also for non-interactive sessions.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>