virtme-ng: restore old --run behavior
Commit d18d080 ("virtme-ng: --run should default to current directory")
changed the behavior of `vng --run` to try to boot a kernel built from
the current working directory, instead of using host's running kernel.

We want to avoid behavior changes, but still give the possibility to
reduce typing as much as possible, especially for the most common use
cases.

Therefore a good compromise between usability and backward compatibility
can be the following:

 - `vng` (no argument): use the kernel from current working directory
 - `vng -r`: use the same kernel installed in the host

In this way both the most common use cases (run a kernel from current
working directory / run the same host's kernel) are easily accessible
with a small amount of typing and, more important, we don't break
backward compatibility.

This fixes issue #32.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>