Use ? on Option where possible
I looked for functions that returned Option like get_user_id() before
this commit. This function calls some other function, check if it
returns a value (Option::Some), and then does something with this value.
If a None was seen, a None is explicitly returned.

This can be simplified with the question mark operator. By using ? on
the Option, the function automatically returns None if a None is seen. I
think this is a bit more readable (although perhaps less beginner
friendly since this might read like magic).

Some functions did similar things but with a Result. Here, I used .ok()?
to get the desired behaviour. I still feel like this is simpler.

Signed-off-by: Uli Schlachter <psychon@znc.in>