Simplify extract_user_script()
This function used find() and manual indexing to split up an input
string. This is all fine, but not terribly obvious to read.

In this commit, I change the code to use split_once() instead. This
finds the first occurrence of some search string and returns the
substrings before and after this. I think this makes the function much
more obvious.

Additionally, I changed end_marker from a string to a char. Both
split_once() and the previously used find() can get a string or a
character as argument. I guess a single character can be slightly more
efficient, but I do not actually know. It just seemed to be the right
thing.

Signed-off-by: Uli Schlachter <psychon@znc.in>