Refactor logic in find_udevd()
The code checks a series of paths for existence. The first one that is
found determines the function return value. First, two static candidates
are checked. Then, for each entry in $PATH, it is checked whether
"udevd" exists in that directory.

This commit refactors that. The if-chain is turned into iterators.
First, the static candidates are put into an array. Then, $PATH is
checked and each entry is mapped to a PathBuf with "udevd" appended.
These two iterators are then chained and the first entry that exists is
returned. This also automatically produces None if no entry exists.

This introduces new allocations for the iterator producing PathBufs.
However, since iterators are lazy, these allocations only actually
happen when the element is needed. Thus, this should be mostly
equivalent to the previous code, except that $PATH is always looked up.

Signed-off-by: Uli Schlachter <psychon@znc.in>