virtme-ng-init: always set XDG_RUNTIME_DIR
XDG_RUNTIME_DIR is not strictly related to X (graphical) applications,
according to [1]:

  XDG_RUNTIME_DIR defines the base directory relative to which
  user-specific non-essential runtime files and other file objects (such
  as sockets, named pipes, ...) should be stored.

Always setting this variable doesn't really introduce any measurable
overhead and it allows to potentially extend the variety of tests and
applications running inside virtme-ng.

Therefore make sure to set this variable and initialize the
XDG_RUNTIME_DIR properly also when we are running in non-graphical mode.

[1] https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html#variables

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>