virtme-ng: introduce argument --root-release
When the target directory used with --root doesn't exist, we create it
from scratch downloading the latest Ubuntu cloud image from the same
release that is running on the host system.

However, virtme-ng can be used also on other distro (not necessarily
Ubuntu), or even on Ubuntu we may want to test a kernel on different
releases.

For this reason add a --root-release option to specify a particular
Ubuntu release as the target to initialize a new root filesystem.

Example (run the latest 24.04 Ubuntu kernel on an Ubuntu 22.04 rootfs):

 $ cat /etc/lsb-release
 DISTRIB_ID=Ubuntu
 DISTRIB_RELEASE=24.04
 DISTRIB_CODENAME=noble
 DISTRIB_DESCRIPTION="Ubuntu Noble Numbat (development branch)"
 $ vng -r --user root --root ./rootfs/22.04 --root-release jammy -- cat /etc/lsb-release
 ...
 DISTRIB_ID=Ubuntu
 DISTRIB_RELEASE=22.04
 DISTRIB_CODENAME=jammy
 DISTRIB_DESCRIPTION="Ubuntu 22.04.3 LTS"

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>