virtme: use kernel modules from the rootfs when --root is used
If we use a separate directory (chroot) as the root filesystem we should
always rely on the content of /lib/modules inside that chroot.

We don't want to try to symlink the modules in the kernel build
directory or any other directory, because we may not be able to reach
that directory from inside the chroot, and this could actually create
problems in some cases (like hiding the modules behind a broken
symlink).

So simply rely on what is available inside the chroot and be happy with
that.

We may improve this in the future by creating additional exports to the
guest (either via 9pfs or virtiofs), but it sounds a bit overkill, so
let's just stick with the simple and most intuitive behavior for now
(that is searching the modules inside the chroot).

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>