virtme: do not enable native features if kvm cannot be used
Do not enable microvm or native kvm acceleration if /dev/kvm does not
exist (or if we don't have permissions to use it).

This may happen in container environments, despite running on a native
architecture, but not having access to /dev/kvm may result in EPERM
errors and warnings that may confuse users.

This fixes issue #51.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>