virtme-ng: introduce NUMA support
Add the --numa option to create custom NUMA nodes inside the guest and
assign virtual CPUs to them.

Syntax:

  --numa MEM[,cpus=FIRST_CPU1[-LAST_CPU1]][,cpus=FIRST_CPU2[-LAST_CPU2]]...

Example:

 $ vng -r -m 4G --numa 2G,cpus=0-1,cpus=3 --numa 2G,cpus=2,cpus=4-7 -- numactl -H
 available: 2 nodes (0-1)
 node 0 cpus: 0 1 3
 node 0 size: 2013 MB
 node 0 free: 1939 MB
 node 1 cpus: 2 4 5 6 7
 node 1 size: 1908 MB
 node 1 free: 1782 MB
 node distances:
 node   0   1
   0:  10  20
   1:  20  10

Keep in mind that using --numa automatically disables the microvm
architecture (that doesn't expose NUMA information).

Also note that virtiofsd requires a memory backed NUMA node, so if some
custom NUMA nodes are created virtiofs will use the first one of them,
otherwise it will create a single NUMA node that maps all the memory
(that was the previous behavior).

But this allows to use --numa and enable virtiofsd at the same time.

This implements the request from issue #56.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>