virtme-ng: channel the return code of a command to the host
Send the return code of the command executed inside the guest to the
host.

This allows to run commands inside a guest and give users the impression
that they are running on the host.

Examples:

  $ vng -r -- true && echo OK || echo ERROR: $?
  OK
  $ vng -r -- false && echo OK || echo ERROR: $?
  ERROR: 1
  $ vng -r -- exit 22 && echo OK || echo ERROR: $?
  ERROR: 22

This feature is particularly useful for automation/CI, since it easily
allows to check the return code of a script executed inside virtme-ng as
if it was running directly on the host.

This fixes issue #48.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>