virtme-configkernel: disable lockdep by default
Commit cd3c35c73624 ("virtme-configkernel: enable lockdep by default")
enabled lockdep, and no doubt lockdep is a useful tool. But it's a bit
surprising that lockdep is enabled on a normal vng build. I would have
expected "vanilla vng build" to be pretty close to kvm_guest.config.
lockdep slows down the kernel, eats memory and is not exactly free of
false positives.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>