virtme-ng: redirect kernel log to stderr in interactive mode
Redirect kernel messages to stderr when running in interactive mode,
similar to what we do in command/script mode.

In this way it is possible, for example, to save kernel logs to a file
even when running in interactive mode, such as:

 $ vng -vr 2>/tmp/kernel.log

NOTE: unfortunately we can't redirect early boot kernel messages,
earlyprintk doesn't seem to work with virconsole devices. So, early
messages will be still sent to stdout, in this way we don't break the
old behavior.

This addresses discussion #60.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>