virtme: use --no-announce-submounts with virtiofsd
Overlayfs doesn't seem to like very much dentries with the flag
DCACHE_NEED_AUTOMOUNT set, see ovl_dentry_weird().

In this case any operation with the dentry would just return EREMOTE.

This seems to happen using virtiofsd on top of filesystems that present
sub-mountpoints. By default virtiofsd announces to the guest which
directories are sub-mountpoints, triggering the problem.

For example, this can be easily reproduced on openSUSE Tumbleweed, that
uses btrfs by default with multile subvolume trees.

Simple reproducer:

 $ vng --disable-microvm -- stat /usr/local
 stat: cannot statx '/usr/local': Object is remote

To avoid this scenario, use the --no-announce-submounts option with
virtiofsd, which prevents the guest from knowing which directories
are sub-mountpoints.

With this change applied the previous test case works as expected:

 $ vng --disable-microvm -- stat /usr/local
   File: /usr/local
   Size: 80        	Blocks: 0          IO Block: 4096   directory
 Device: 0,41	Inode: 10          Links: 1
 Access: (0755/drwxr-xr-x)  Uid: (    0/    root)   Gid: (    0/    root)
 Access: 2024-02-15 14:24:16.406585004 +0100
 Modify: 2024-02-15 10:39:00.030001193 +0100
 Change: 2024-02-15 10:39:00.030001193 +0100
  Birth: -

This fixes issue #50.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>